test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

# Hacking already pins down pep8, pyflakes and flake8
hacking>=0.10.0,<0.11
Babel>=1.3
contextlib2>=0.4.0 # PSF License
coverage>=3.6
discover
fixtures>=0.3.14
happybase>=0.5,!=0.7
httplib2>=0.7.5
mock>=1.0
PyMySQL>=0.6.2  # MIT License
# Docs Requirements
oslosphinx>=2.5.0  # Apache-2.0
oslotest>=1.5.1  # Apache-2.0
oslo.vmware>=0.13.1                     # Apache-2.0
psycopg2
pylint==1.4.1  # GNU GPL v2
pymongo>=3.0.2
python-subunit>=0.0.18
sphinx>=1.1.2,!=1.2.0,!=1.3b1,<1.3
sphinxcontrib-docbookrestapi
sphinxcontrib-httpdomain
sphinxcontrib-pecanwsme>=0.8
testrepository>=0.0.18
testscenarios>=0.4
testtools>=0.9.36,!=1.2.0
gabbi>=0.12.0  # Apache-2.0
requests-aws>=0.1.4    # BSD License (3 clause)
tempest-lib>=0.5.0
